Manage cache / download directories for study data

Managing data downloads is important to save disk space and re-downloading data files. This can be done effortlessly via the integrated BiocFileCache system.

cBioCache(...)

setCache(
  directory = rappdirs::user_cache_dir("cBioPortalData"),
  verbose = TRUE,
  ask = interactive()
)

removeCache(cancer_study_id)

Arguments

...
For cBioCache, arguments passed to setCache

directory
The file location where the cache is located. Once set future downloads will go to this folder.

verbose
Whether to print descriptive messages

ask
logical (default TRUE when interactive session) Confirm the file location of the cache directory

cancer_study_id
A single string from studiesTable associated with a study tarball

Value

cBioCache: The path to the cache location

cBioCache

Get the directory location of the cache. It will prompt the user to create a cache if not already created. A specific directory can be used via setCache.

setCache

Specify the directory location of the data cache. By default, it will got to the user's home/.cache and "appname" directory as specified by user_cache_dir. (default appname: cBioPortalData)

removeCache

Some files may become corrupt when downloading, this function allows the user to delete the tarball associated with a cancer_study_id in the cache. This only works for the cBioDataPack function. To remove the entire cBioPortalData cache, run unlink("~/.cache/cBioPortalData").

Examples

(cacheloc <- cBioCache())

#> [1] "/home/mr148/.cache/cBioPortalData"