Obtain pre-packaged data from cBioPortal and represent as a MultiAssayExperiment object

The cBioDataPack function allows the user to download and process cancer study datasets found in MSKCC's cBioPortal. Output datasets use the MultiAssayExperiment data representation to faciliate analysis and data management operations.

cBioDataPack(
  cancer_study_id,
  use_cache = TRUE,
  split.field = c("Tumor_Sample_Barcode", "ID"),
  names.field = c("Hugo_Symbol", "Entrez_Gene_Id", "Gene")
)

Arguments

cancer_study_id
The cBioPortal study identifier as in https://cbioportal.org/webAPI

use_cache
logical (default TRUE) create the default cache location and use it to track downloaded data. If data found in the cache, data will not be re-downloaded. A path can also be provided to data cache location.

split.field
A character vector of possible column names for the column that is used to identify samples in a mutations or copy number file.

names.field
A character vector of possible column names for the column that is used to label ranges from a mutations or copy number file.

Value

A MultiAssayExperiment object

Details

The list of datasets can be found in the `studiesTable` dataset by doing data("studiesTable"). Some datasets may not be available for download. Please refer to the website for the full list of datasets

See also

Examples

data(studiesTable)

head(studiesTable[["cancer_study_id"]])

#> [1] "paac_jhu_2014"       "mel_tsam_liang_2017" "all_stjude_2015"    
#> [4] "all_stjude_2016"     "aml_ohsu_2018"       "laml_tcga_pub"      

mae <- cBioDataPack("laml_tcga")

#> Study file in cache: laml_tcga

#> Working on: /tmp/Rtmp7hNsIM/data_CNA.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmp7hNsIM/data_RNA_Seq_expression_median.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmp7hNsIM/data_RNA_Seq_mRNA_median_Zscores.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmp7hNsIM/data_RNA_Seq_v2_expression_median.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmp7hNsIM/data_RNA_Seq_v2_mRNA_median_Zscores.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmp7hNsIM/data_cna_hg19.seg

#> Parsed with column specification:
#> cols(
#>   ID = col_character(),
#>   chrom = col_double(),
#>   loc.start = col_double(),
#>   loc.end = col_double(),
#>   num.mark = col_double(),
#>   seg.mean = col_double()
#> )

#> Working on: /tmp/Rtmp7hNsIM/data_linear_CNA.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmp7hNsIM/data_methylation_hm27.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmp7hNsIM/data_methylation_hm450.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmp7hNsIM/data_mutations_extended.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Entrez_Gene_Id = col_double(),
#>   Start_Position = col_double(),
#>   End_Position = col_double(),
#>   dbSNP_Val_Status = col_logical(),
#>   Score = col_double(),
#>   t_ref_count = col_double(),
#>   t_alt_count = col_double(),
#>   n_ref_count = col_double(),
#>   n_alt_count = col_double(),
#>   Protein_position = col_double(),
#>   Hotspot = col_double(),
#>   RNAVAF_WU = col_double(),
#>   RNAVarReads_WU = col_double(),
#>   stop = col_double(),
#>   NormalVAF_WU = col_double(),
#>   start = col_double(),
#>   TumorVAF_WU = col_double(),
#>   RNARefReads_WU = col_double()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character()
#> )

#> See spec(...) for full column specifications.

#> Warning: cannnot determine seqnames column unambiguously

#> Warning: cannnot determine seqnames column unambiguously

#> Warning: cannnot determine seqnames column unambiguously

#> Working on: /tmp/Rtmp7hNsIM/data_mutations_mskcc.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Entrez_Gene_Id = col_double(),
#>   Start_Position = col_double(),
#>   End_Position = col_double(),
#>   dbSNP_Val_Status = col_logical(),
#>   Score = col_double(),
#>   t_ref_count = col_double(),
#>   t_alt_count = col_double(),
#>   n_ref_count = col_double(),
#>   n_alt_count = col_double(),
#>   Protein_position = col_double(),
#>   Hotspot = col_double(),
#>   RNAVAF_WU = col_double(),
#>   RNAVarReads_WU = col_double(),
#>   stop = col_double(),
#>   NormalVAF_WU = col_double(),
#>   start = col_double(),
#>   TumorVAF_WU = col_double(),
#>   RNARefReads_WU = col_double()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character()
#> )

#> See spec(...) for full column specifications.

#> Warning: cannnot determine seqnames column unambiguously

#> Warning: cannnot determine seqnames column unambiguously

#> Warning: cannnot determine seqnames column unambiguously

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   INITIAL_PATHOLOGIC_DX_YEAR = col_double(),
#>   AGE = col_double(),
#>   PLATELET_COUNT_PRERESECTION = col_double(),
#>   BLAST_COUNT = col_double(),
#>   BASOPHILS_COUNT = col_double(),
#>   ABNORMAL_LYMPHOCYTE_PERCENT = col_double(),
#>   OS_MONTHS = col_double()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   SAMPLE_TYPE_ID = col_double()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   SAMPLE_TYPE_ID = col_double()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   index = col_double(),
#>   chromosome = col_double(),
#>   region_start = col_double(),
#>   region_end = col_double(),
#>   peak_start = col_double(),
#>   peak_end = col_double(),
#>   enlarged_peak_start = col_double(),
#>   enlarged_peak_end = col_double(),
#>   n_genes_in_region = col_double(),
#>   genes_in_region = col_character(),
#>   n_genes_in_peak = col_double(),
#>   genes_in_peak = col_character(),
#>   n_genes_on_chip = col_character(),
#>   genes_on_chip = col_character(),
#>   `top 3` = col_character(),
#>   amp = col_double(),
#>   cytoband = col_character(),
#>   q_value = col_double()
#> )

#> Warning:  Multiple prefixes found, using keyword 'region' or taking first one

#> Parsed with column specification:
#> cols(
#>   index = col_double(),
#>   chromosome = col_double(),
#>   region_start = col_double(),
#>   region_end = col_double(),
#>   peak_start = col_double(),
#>   peak_end = col_double(),
#>   enlarged_peak_start = col_double(),
#>   enlarged_peak_end = col_double(),
#>   n_genes_in_region = col_double(),
#>   genes_in_region = col_character(),
#>   n_genes_in_peak = col_double(),
#>   genes_in_peak = col_character(),
#>   n_genes_on_chip = col_character(),
#>   genes_on_chip = col_character(),
#>   `top 3` = col_character(),
#>   amp = col_double(),
#>   cytoband = col_character(),
#>   q_value = col_double()
#> )

#> Warning:  Multiple prefixes found, using keyword 'region' or taking first one