The R interface to the cBioPortal API Data Service

This function gives users access to the cBioPortal data with an API representation invoked from `cBioPortal()`

cBioPortal()

getStudies(api)

clinicalData(api, studyId = NA_character_)

molecularProfiles(
  api,
  studyId = NA_character_,
  projection = c("SUMMARY", "ID", "DETAILED", "META")
)

molecularData(
  api,
  molecularProfileId = NA_character_,
  entrezGeneIds = NULL,
  sampleIds = NULL,
  check = TRUE
)

searchOps(api, keyword)

geneTable(api, ...)

samplesInSampleLists(api, sampleListIds = NA_character_, check = TRUE)

sampleLists(api, studyId = NA_character_)

allSamples(api, studyId = NA_character_)

genePanels(api)

getGenePanel(api, genePanelId = NA_character_)

genePanelMolecular(
  api,
  molecularProfileId = NA_character_,
  sampleListId = NULL,
  sampleIds = NULL
)

getGenePanelMolecular(api, molecularProfileIds = NA_character_, sampleIds)

getSampleInfo(
  api,
  studyId = NA_character_,
  sampleListIds = NULL,
  projection = c("SUMMARY", "ID", "DETAILED", "META")
)

getDataByGenePanel(
  api,
  studyId = NA_character_,
  genePanelId = NA_character_,
  molecularProfileId = NULL,
  sampleListId = NULL,
  check = TRUE
)

Arguments

api
An API object of class `cBioPortal` from the `cBioPortal` function

studyId
character(1) Indicates the "studyId" as taken from `getStudies`

projection
character(default: "SUMMARY") Specify the projection type for data retrieval for details see API documentation

molecularProfileId
character(1) Indicates a molecular profile ID

entrezGeneIds
numeric() A vector indicating entrez gene IDs

sampleIds
character() Sample identifiers

check
logical(1) Whether to check the inputs against values from the API (i.e., for 'studyId', 'genePanelId', 'molecularProfileId', and 'sampleListId')

keyword
character(1) Keyword or pattern for searching through available operations

...
Additional arguments to lower level API functions

sampleListIds
character() A vector of 'sampleListId' as obtained from `sampleLists`

genePanelId
character(1) Identifies the gene panel, as obtained from the `genePanels` function

sampleListId
character(1) A sample list identifier as obtained from `sampleLists()``

molecularProfileIds
character() A vector of molecular profile IDs

Value

cBioPortal: An API object of class 'cBioPortal'

cBioPortalData: A data object of class 'MultiAssayExperiment'

API Metadata

* getStudies - Obtain a table of studies and associated metadata

* searchOps - Search through API operations with a keyword

* geneTable - Get a table of all genes by 'entrezGeneId' or 'hugoGeneSymbol'

* sampleLists - obtain all `sampleListIds` for a particular `studyId`

* allSamples - obtain all samples within a particular `studyId`

* genePanels - Show all available gene panels

Patient Data

* clinicalData - Obtain clinical data for a particular study identifier ('studyId')

Molecular Profiles

* molecularProfiles - Produce a molecular profiles dataset for a given study identifier ('studyId')

* molecularData - Produce a dataset of molecular profile data based on `molecularProfileId`, `entrezGeneIds`, and `sampleIds`

Sample Data

* samplesInSampleLists - get all samples associated with a 'sampleListId'

* getSampleInfo - Obtain sample metadata for a particular `studyId` or `sampleListId`

Gene Panels

* getGenePanels - Obtain the gene panel for a particular 'genePanelId'

* genePanelMolecular - get gene panel data for a paricular `molecularProfileId` and `sampleListId` combination

* getGenePanelMolecular - get gene panel data for a combination of `molecularProfileId` and `sampleListId` vectors

* getDataByGenePanel - Download data for a gene panel and `molecularProfileId` combination, optionally a `sampleListId` can be provided.

Examples

cbio <- cBioPortal()

getStudies(api = cbio)

#> # A tibble: 280 x 13
#>    name  shortName description publicStudy pmid  citation groups status
#>    <chr> <chr>     <chr>       <lgl>       <chr> <chr>    <chr>   <int>
#>  1 Chol… Cholangi… Exome sequ… TRUE        2418… Chan-on… "PUBL…      0
#>  2 Chol… Cholangi… Exome Sequ… TRUE        2256… Ong et … "PUBL…      0
#>  3 Cuta… CTCL (Co… Whole-Exom… TRUE        2655… Da Silv… ""          0
#>  4 Esop… ESCC (UC… Whole exom… TRUE        2468… Lin et … "PUBL…      0
#>  5 Oral… Head & n… Comprehens… TRUE        2361… Pickeri… ""          0
#>  6 Hepa… HCC (Ins… Whole-exom… TRUE        2582… Schulze… "PUBL…      0
#>  7 Uvea… UM (QIMR) Whole-geno… TRUE        2668… Johanss… "PUBL…      0
#>  8 Neur… NBL (AMC) Whole geno… TRUE        2236… Molenaa… "PUBL…      0
#>  9 Naso… NPC (Sin… Whole exom… TRUE        2495… Lin et … "PUBL…      0
#> 10 Thym… TET (NCI) Whole exom… TRUE        2497… Petrini… "PUBL…      0
#> # … with 270 more rows, and 5 more variables: importDate <chr>,
#> #   allSampleCount <int>, studyId <chr>, cancerTypeId <chr>,
#> #   referenceGenome <chr>

searchOps(api = cbio, keyword = "molecular")

#>  [1] "fetchGenePanelDataInMultipleMolecularProfilesUsingPOST"   
#>  [2] "fetchGenericAssayDataInMultipleMolecularProfilesUsingPOST"
#>  [3] "fetchGenericAssayDataInMolecularProfileUsingPOST"         
#>  [4] "fetchMolecularDataInMultipleMolecularProfilesUsingPOST"   
#>  [5] "getAllMolecularProfilesUsingGET"                          
#>  [6] "fetchMolecularProfilesUsingPOST"                          
#>  [7] "getMolecularProfileUsingGET"                              
#>  [8] "getDiscreteCopyNumbersInMolecularProfileUsingGET"         
#>  [9] "fetchDiscreteCopyNumbersInMolecularProfileUsingPOST"      
#> [10] "getAllMolecularDataInMolecularProfileUsingGET"            
#> [11] "fetchAllMolecularDataInMolecularProfileUsingPOST"         
#> [12] "getMutationsInMolecularProfileBySampleListIdUsingGET"     
#> [13] "fetchMutationsInMolecularProfileUsingPOST"                
#> [14] "fetchMutationsInMultipleMolecularProfilesUsingPOST"       
#> [15] "getAllMolecularProfilesInStudyUsingGET"                   

clinicalData(api = cbio, studyId = "acc_tcga")

#> # A tibble: 92 x 70
#>    uniquePatientKey patientId studyId AGE   AJCC_PATHOLOGIC… ATYPICAL_MITOTI…
#>    <chr>            <chr>     <chr>   <chr> <chr>            <chr>           
#>  1 VENHQS1PUi1BNUo… TCGA-OR-… acc_tc… 23    Stage IV         Atypical Mitoti…
#>  2 VENHQS1PUi1BNUo… TCGA-OR-… acc_tc… 30    Stage III        Atypical Mitoti…
#>  3 VENHQS1PUi1BNUo… TCGA-OR-… acc_tc… 29    Stage II         Atypical Mitoti…
#>  4 VENHQS1PUi1BNUo… TCGA-OR-… acc_tc… 30    Stage III        Atypical Mitoti…
#>  5 VENHQS1PUi1BNUo… TCGA-OR-… acc_tc… 66    Stage III        Atypical Mitoti…
#>  6 VENHQS1PUi1BNUo… TCGA-OR-… acc_tc… 22    Stage II         Atypical Mitoti…
#>  7 VENHQS1PUi1BNUo… TCGA-OR-… acc_tc… 58    Stage II         Atypical Mitoti…
#>  8 VENHQS1PUi1BNUo… TCGA-OR-… acc_tc… 44    Stage IV         Atypical Mitoti…
#>  9 VENHQS1PUi1BNUo… TCGA-OR-… acc_tc… 23    Stage III        Atypical Mitoti…
#> 10 VENHQS1PUi1BNUp… TCGA-OR-… acc_tc… 60    Stage II         NA              
#> # … with 82 more rows, and 64 more variables: CAPSULAR_INVASION <chr>,
#> #   CLIN_M_STAGE <chr>, CLINICAL_STATUS_WITHIN_3_MTHS_SURGERY <chr>,
#> #   CT_SCAN_PREOP_RESULTS <chr>,
#> #   CYTOPLASM_PRESENCE_LESS_THAN_EQUAL_25_PERCENT <chr>,
#> #   DAYS_LAST_FOLLOWUP <chr>, DAYS_TO_INITIAL_PATHOLOGIC_DIAGNOSIS <chr>,
#> #   DFS_MONTHS <chr>, DFS_STATUS <chr>, DIFFUSE_ARCHITECTURE <chr>,
#> #   ETHNICITY <chr>, FORM_COMPLETION_DATE <chr>, HISTOLOGICAL_DIAGNOSIS <chr>,
#> #   HISTORY_ADRENAL_HORMONE_EXCESS <chr>,
#> #   HISTORY_BASIS_ADRENAL_HORMONE_DX <chr>, HISTORY_NEOADJUVANT_TRTYN <chr>,
#> #   HISTORY_OTHER_MALIGNANCY <chr>, ICD_10 <chr>, ICD_O_3_HISTOLOGY <chr>,
#> #   ICD_O_3_SITE <chr>, INFORMED_CONSENT_VERIFIED <chr>,
#> #   INITIAL_PATHOLOGIC_DX_YEAR <chr>, LATERALITY <chr>,
#> #   LYMPH_NODE_EXAMINED_COUNT <chr>, LYMPH_NODES_EXAMINED <chr>,
#> #   LYMPH_NODES_EXAMINED_HE_COUNT <chr>, METASTATIC_DX_CONFIRMED_BY <chr>,
#> #   METASTATIC_DX_CONFIRMED_BY_OTHER <chr>, METASTATIC_SITE_PATIENT <chr>,
#> #   MITOSES_PER_50_HPF <chr>, MITOTIC_RATE <chr>,
#> #   MOLECULAR_STUDIES_OTHERS_PERFORMED <chr>, NECROSIS <chr>,
#> #   NEW_TUMOR_EVENT_AFTER_INITIAL_TREATMENT <chr>, NUCLEAR_GRADE_III_IV <chr>,
#> #   OS_MONTHS <chr>, OS_STATUS <chr>, OTHER_PATIENT_ID <chr>,
#> #   PATH_N_STAGE <chr>, PATH_T_STAGE <chr>, PHARM_TX_MITOTANE_ADJUVANT <chr>,
#> #   PHARM_TX_MITOTANE_FOR_MACRO_DISEASE <chr>,
#> #   PHARM_TX_MITOTANE_INDICATOR <chr>,
#> #   PHARM_TX_MITOTANE_THERAPUTIC_AT_PROG <chr>,
#> #   PHARM_TX_MITOTANE_THERAPUTIC_AT_REC <chr>,
#> #   PHARM_TX_MITOTANE_THERAPUTIC_LEVELS <chr>,
#> #   PHARM_TX_MITOTANE_THERAPUTIC_MACRO <chr>, PHARMACEUTICAL_TX_ADJUVANT <chr>,
#> #   PROSPECTIVE_COLLECTION <chr>, RACE <chr>,
#> #   RADIATION_TREATMENT_ADJUVANT <chr>, RESIDUAL_TUMOR <chr>, RET <chr>,
#> #   RETROSPECTIVE_COLLECTION <chr>, SAMPLE_COUNT <chr>, SEX <chr>,
#> #   SINUSOID_INVASION <chr>, SITE_OF_TUMOR_TISSUE <chr>,
#> #   TISSUE_SOURCE_SITE <chr>, TREATMENT_OUTCOME_FIRST_COURSE <chr>,
#> #   TUMOR_STATUS <chr>, VITAL_STATUS <chr>, WEISS_SCORE_OVERALL <chr>,
#> #   WEISS_VENOUS_INVASION <chr>

molecularProfiles(api = cbio, studyId = "acc_tcga")

#> # A tibble: 9 x 7
#>   molecularAltera… datatype name  description showProfileInAn… molecularProfil…
#>   <chr>            <chr>    <chr> <chr>       <lgl>            <chr>           
#> 1 PROTEIN_LEVEL    LOG2-VA… Prot… Protein ex… FALSE            acc_tcga_rppa   
#> 2 PROTEIN_LEVEL    Z-SCORE  Prot… Protein ex… TRUE             acc_tcga_rppa_Z…
#> 3 COPY_NUMBER_ALT… DISCRETE Puta… Putative c… TRUE             acc_tcga_gistic 
#> 4 MRNA_EXPRESSION  CONTINU… mRNA… mRNA gene … FALSE            acc_tcga_rna_se…
#> 5 MRNA_EXPRESSION  Z-SCORE  mRNA… mRNA expre… TRUE             acc_tcga_rna_se…
#> 6 COPY_NUMBER_ALT… CONTINU… Rela… Relative l… FALSE            acc_tcga_linear…
#> 7 METHYLATION      CONTINU… Meth… Methylatio… FALSE            acc_tcga_methyl…
#> 8 MUTATION_EXTEND… MAF      Muta… Mutation d… TRUE             acc_tcga_mutati…
#> 9 MRNA_EXPRESSION  Z-SCORE  mRNA… Log-transf… TRUE             acc_tcga_rna_se…
#> # … with 1 more variable: studyId <chr>

molecularData(
    api = cbio,
    molecularProfileId = "acc_tcga_rna_seq_v2_mrna",
    entrezGeneIds = c(1, 2),
    sampleIds = c("TCGA-OR-A5J1-01", "TCGA-OR-A5J2-01")
)

#> # A tibble: 4 x 8
#>   uniqueSampleKey uniquePatientKey entrezGeneId molecularProfil… sampleId
#>   <chr>           <chr>                   <int> <chr>            <chr>   
#> 1 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…            1 acc_tcga_rna_se… TCGA-OR…
#> 2 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…            2 acc_tcga_rna_se… TCGA-OR…
#> 3 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…            1 acc_tcga_rna_se… TCGA-OR…
#> 4 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…            2 acc_tcga_rna_se… TCGA-OR…
#> # … with 3 more variables: patientId <chr>, studyId <chr>, value <dbl>

sampleLists(api = cbio, studyId = "acc_tcga")

#> # A tibble: 9 x 5
#>   category          name          description            sampleListId    studyId
#>   <chr>             <chr>         <chr>                  <chr>           <chr>  
#> 1 all_cases_with_r… Samples with… Samples protein data … acc_tcga_rppa   acc_tc…
#> 2 all_cases_with_m… Samples with… Samples with mutation… acc_tcga_cnaseq acc_tc…
#> 3 all_cases_in_stu… All samples   All samples (92 sampl… acc_tcga_all    acc_tc…
#> 4 all_cases_with_c… Samples with… Samples with CNA data… acc_tcga_cna    acc_tc…
#> 5 all_cases_with_m… Samples with… Samples with mutation… acc_tcga_seque… acc_tc…
#> 6 all_cases_with_m… Samples with… Samples with methylat… acc_tcga_methy… acc_tc…
#> 7 all_cases_with_m… Samples with… Samples with mRNA exp… acc_tcga_rna_s… acc_tc…
#> 8 all_cases_with_m… Complete sam… Samples with mutation… acc_tcga_3way_… acc_tc…
#> 9 all_cases_with_m… Samples with… Samples with methylat… acc_tcga_methy… acc_tc…

samplesInSampleLists(
    api = cbio,
    sampleListIds = c("acc_tcga_rppa", "acc_tcga_cnaseq")
)

#> CharacterList of length 2
#> [["acc_tcga_cnaseq"]] TCGA-OR-A5J1-01 TCGA-OR-A5J2-01 ... TCGA-PK-A5HC-01
#> [["acc_tcga_rppa"]] TCGA-OR-A5J2-01 TCGA-OR-A5J3-01 ... TCGA-PK-A5HA-01

genePanels(api = cbio)

#> # A tibble: 46 x 2
#>    description                                                       genePanelId
#>    <chr>                                                             <chr>      
#>  1 Targeted (341 cancer genes) sequencing of various tumor types vi… IMPACT341  
#>  2 Targeted (410 cancer genes) sequencing of various tumor types vi… IMPACT410  
#>  3 Targeted (468 cancer genes) sequencing of various tumor types vi… IMPACT468  
#>  4 Targeted sequencing of urcc tumor via MSK-IMPACT.                 IMPACT     
#>  5 Targeted (300 cancer genes) sequencing of bladder urothelial car… IMPACT300  
#>  6 Targeted deep sequencing (TDS) of 1202 genes on Illumina HiSeq20… UCLA_1202  
#>  7 Targeted sequencing of urcc tumor via MSK-IMPACT.                 IMPACT230  
#>  8 Targeted (27 cancer genes) sequencing of adenoid cystic carcinom… ACYC_FMI_27
#>  9 Targeted (173 cancer genes) sequencing of breast cancers on Illu… METABRIC_1…
#> 10 Targeted sequencing of 504 cancer-associated genes on Illumina H… DFCI_504   
#> # … with 36 more rows

getGenePanel(api = cbio, genePanelId = "IMPACT341")

#> # A tibble: 341 x 2
#>    entrezGeneId hugoGeneSymbol
#>           <int> <chr>         
#>  1           25 ABL1          
#>  2        84142 ABRAXAS1      
#>  3          207 AKT1          
#>  4          208 AKT2          
#>  5        10000 AKT3          
#>  6          238 ALK           
#>  7          242 ALOX12B       
#>  8       139285 AMER1         
#>  9          324 APC           
#> 10          367 AR            
#> # … with 331 more rows

getDataByGenePanel(cbio, studyId = "acc_tcga", genePanelId = "IMPACT341",
   molecularProfileId = "acc_tcga_rppa")

#> # A tibble: 2,622 x 8
#>    uniqueSampleKey uniquePatientKey entrezGeneId molecularProfil… sampleId
#>    <chr>           <chr>                   <int> <chr>            <chr>   
#>  1 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…          207 acc_tcga_rppa    TCGA-OR…
#>  2 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…          208 acc_tcga_rppa    TCGA-OR…
#>  3 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…        10000 acc_tcga_rppa    TCGA-OR…
#>  4 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…          367 acc_tcga_rppa    TCGA-OR…
#>  5 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…          472 acc_tcga_rppa    TCGA-OR…
#>  6 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…         8314 acc_tcga_rppa    TCGA-OR…
#>  7 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…          596 acc_tcga_rppa    TCGA-OR…
#>  8 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…          598 acc_tcga_rppa    TCGA-OR…
#>  9 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…        10018 acc_tcga_rppa    TCGA-OR…
#> 10 VENHQS1PUi1BNU… VENHQS1PUi1BNUo…          673 acc_tcga_rppa    TCGA-OR…
#> # … with 2,612 more rows, and 3 more variables: patientId <chr>, studyId <chr>,
#> #   value <dbl>