A class for representing the cBioPortal API protocol

The `cBioPortal` class is a representation of the cBioPortal API protocol that directly inherits from the `Service` class in the `AnVIL` package. For more information, see the 'AnVIL' package.

Details

This class takes the static API as provided at https://www.cbioportal.org/api/api-docs and creates an R object with the help from underlying infrastructure (i.e., 'rapiclient' and 'AnVIL') to give the user a unified representation of the API specification provided by the cBioPortal group. Users are not expected to interact with this class other than to use it as input to the functionality provided by the rest of the package.

See also

Examples

cBioPortal()

#> service: cBioPortal
#> tags(); use cbioportal$<tab completion>:
#> # A tibble: 73 x 3
#>    tag          operation               summary                                 
#>    <chr>        <chr>                   <chr>                                   
#>  1 Cancer Types getAllCancerTypesUsing… Get all cancer types                    
#>  2 Cancer Types getCancerTypeUsingGET   Get a cancer type                       
#>  3 Clinical At… fetchClinicalAttribute… Fetch clinical attributes               
#>  4 Clinical At… getAllClinicalAttribut… Get all clinical attributes in the spec…
#>  5 Clinical At… getAllClinicalAttribut… Get all clinical attributes             
#>  6 Clinical At… getClinicalAttributeCo… Get counts for clinical attributes acco…
#>  7 Clinical At… getClinicalAttributeIn… Get specified clinical attribute        
#>  8 Clinical Da… fetchAllClinicalDataIn… Fetch clinical data by patient IDs or s…
#>  9 Clinical Da… fetchClinicalDataUsing… Fetch clinical data by patient IDs or s…
#> 10 Clinical Da… getAllClinicalDataInSt… Get all clinical data in a study        
#> # … with 63 more rows
#> tag values:
#>   Cancer Types, Clinical Attributes, Clinical Data, Clinical Events,
#>   Copy Number Segments, Discrete Copy Number Alterations, Gene Panels,
#>   Generic Assays, Genes, Molecular Data, Molecular Profiles, Mutations,
#>   Patients, Reference Genome Genes, Resource Data, Resource
#>   Definitions, Sample Lists, Samples, Structural Variants, Studies,
#>   Treatments
#> schemas():
#>   AlleleSpecificCopyNumber, AndedPatientTreatmentFilters,
#>   AndedSampleTreatmentFilters, CancerStudy, CancerStudyTags
#>   # ... with 66 more elements