Download data from the cBioPortal API

Obtain a MultiAssayExperiment object for a particular gene panel, studyId, molecularProfileIds, and sampleListIds combination. Default molecularProfileIds and sampleListIds are set to NULL for including all data. This option is best for users who wish to obtain a section of the study data that pertains to a specific molecular profile and gene panel combination. For users looking to download the entire study data as provided by the https://cbioportal.org/datasets, refer to cBioDataPack.

cBioPortalData(
  api,
  studyId = NA_character_,
  genePanelId = NA_character_,
  molecularProfileIds = NULL,
  sampleListId = NULL,
  by = c("entrezGeneId", "hugoGeneSymbol")
)

Arguments

api
An API object of class `cBioPortal` from the `cBioPortal` function

studyId
character(1) Indicates the "studyId" as taken from `getStudies`

genePanelId
character(1) Identifies the gene panel, as obtained from the `genePanels` function

molecularProfileIds
character() A vector of molecular profile IDs

sampleListId
character(1) A sample list identifier as obtained from `sampleLists()``

by
character(1) Either 'entrezGeneId' or 'hugoGeneSymbol' for row metadata

Value

A MultiAssayExperiment object

See also

Examples

cbio <- cBioPortal()

cBioPortalData(cbio, by = "hugoGeneSymbol", studyId = "acc_tcga",
    genePanelId = "IMPACT341",
    molecularProfileIds = c("acc_tcga_rppa", "acc_tcga_linear_CNA")
    )

#> Error in is(api, "cBioPortal"): object 'cbio' not found