Manually download, untar, and load study tarballs

Note that these functions should be used when a particular study is not currently available as a MultiAssayExperiment representation. Otherwise, use cBioDataPack. Provide a cancer_study_id from the studiesTable and retrieve the study tarball from cBioPortal. These functions are used by cBioDataPack under the hood to download, untar, and load the tarball datasets with caching. As stated in ?cBioDataPack, not all studies are currently working as MultiAssayExperiment objects. As of July 2020, about ~80% of datasets can be successfully imported into the MultiAssayExperiment data class. Please open an issue if you would like the team to prioritize a study. You may also check studiesTable$pack_build for a more current status.

downloadStudy(
  cancer_study_id,
  use_cache = TRUE,
  force = FALSE,
  url_location = getOption("cBio_URL", .url_location)
)

untarStudy(cancer_study_file, exdir = tempdir())

loadStudy(filepath, names.field = c("Hugo_Symbol", "Entrez_Gene_Id", "Gene"))

Arguments

cancer_study_id
character(1) The study identifier from cBioPortal as in https://cbioportal.org/webAPI

use_cache
logical(1) (default TRUE) create the default cache location and use it to track downloaded data. If data found in the cache, data will not be re-downloaded. A path can also be provided to data cache location.

force
logical(1) (default FALSE) whether to force re-download data from remote location

url_location
character(1) (default "https://cbioportal-datahub.s3.amazonaws.com") the URL location for downloading packaged data. Can be set using the 'cBio_URL' option (see ?cBioDataPack for more details)

cancer_study_file
character(1) indicates the on-disk location of the downloaded tarball

exdir
character(1) indicates the folder location to put the contents of the tarball (default tempdir(); see also ?untar)

filepath
character(1) indicates the folder location where the contents of the tarball are located (usually the same as exdir)

names.field
A character vector of possible column names for the column that is used to label ranges from a mutations or copy number file.

Value

downloadStudy: The file location of the data tarball
untarStudy: The directory location of the contents
loadStudy: A \linkS4class{MultiAssayExperiment} object

See also

Examples

(acc_file <- downloadStudy("acc_tcga"))

#> Study file in cache: acc_tcga

#>                                                                BFC6 
#> "/github/home/.cache/R/cBioPortalData/7784170df381_acc_tcga.tar.gz" 

(file_dir <- untarStudy(acc_file, tempdir()))

#> [1] "/tmp/Rtmpphincs/7784170df381_acc_tcga"

loadStudy(file_dir)

#> Working on: /tmp/Rtmpphincs/7784170df381_acc_tcga/data_cna_hg19.seg

#> Parsed with column specification:
#> cols(
#>   ID = col_character(),
#>   chrom = col_double(),
#>   loc.start = col_double(),
#>   loc.end = col_double(),
#>   num.mark = col_double(),
#>   seg.mean = col_double()
#> )

#> Working on: /tmp/Rtmpphincs/7784170df381_acc_tcga/data_CNA.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmpphincs/7784170df381_acc_tcga/data_linear_CNA.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmpphincs/7784170df381_acc_tcga/data_methylation_hm450.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmpphincs/7784170df381_acc_tcga/data_mutations_extended.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Entrez_Gene_Id = col_double(),
#>   Start_Position = col_double(),
#>   End_Position = col_double(),
#>   t_ref_count = col_logical(),
#>   t_alt_count = col_logical(),
#>   n_ref_count = col_logical(),
#>   n_alt_count = col_logical(),
#>   Protein_position = col_double(),
#>   Hotspot = col_double(),
#>   TVarCov = col_double(),
#>   Trna_ref = col_double(),
#>   NTotCov = col_double(),
#>   Trna_var = col_double(),
#>   NVarCov = col_double(),
#>   Trna_alt1 = col_double(),
#>   Trna_alt2 = col_double(),
#>   Trna_tot = col_double(),
#>   TTotCov = col_double()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character()
#> )

#> See spec(...) for full column specifications.

#> Working on: /tmp/Rtmpphincs/7784170df381_acc_tcga/data_mutations_mskcc.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Entrez_Gene_Id = col_double(),
#>   Start_Position = col_double(),
#>   End_Position = col_double(),
#>   t_ref_count = col_logical(),
#>   t_alt_count = col_logical(),
#>   n_ref_count = col_logical(),
#>   n_alt_count = col_logical(),
#>   Protein_position = col_double(),
#>   Hotspot = col_double(),
#>   TVarCov = col_double(),
#>   Trna_ref = col_double(),
#>   NTotCov = col_double(),
#>   Trna_var = col_double(),
#>   NVarCov = col_double(),
#>   Trna_alt1 = col_double(),
#>   Trna_alt2 = col_double(),
#>   Trna_tot = col_double(),
#>   TTotCov = col_double()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character()
#> )

#> See spec(...) for full column specifications.

#> Working on: /tmp/Rtmpphincs/7784170df381_acc_tcga/data_mutsig.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   gene = col_character(),
#>   description = col_character(),
#>   p = col_character(),
#>   q = col_character()
#> )

#> See spec(...) for full column specifications.

#> Working on: /tmp/Rtmpphincs/7784170df381_acc_tcga/data_RNA_Seq_v2_expression_median.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmpphincs/7784170df381_acc_tcga/data_RNA_Seq_v2_mRNA_median_all_sample_Zscores.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmpphincs/7784170df381_acc_tcga/data_RNA_Seq_v2_mRNA_median_Zscores.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Hugo_Symbol = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Hugo_Symbol = col_character()
#> )

#> Working on: /tmp/Rtmpphincs/7784170df381_acc_tcga/data_rppa_Zscores.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Composite.Element.REF = col_character()
#> )

#> See spec(...) for full column specifications.

#> Working on: /tmp/Rtmpphincs/7784170df381_acc_tcga/data_rppa.txt

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   Composite.Element.REF = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   INITIAL_PATHOLOGIC_DX_YEAR = col_double(),
#>   WEISS_SCORE_OVERALL = col_double(),
#>   MITOSES_PER_50_HPF = col_double(),
#>   AGE = col_double(),
#>   DAYS_TO_INITIAL_PATHOLOGIC_DIAGNOSIS = col_double(),
#>   OS_MONTHS = col_double(),
#>   DFS_MONTHS = col_double()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   DAYS_TO_COLLECTION = col_double(),
#>   SAMPLE_INITIAL_WEIGHT = col_double(),
#>   OCT_EMBEDDED = col_logical(),
#>   SAMPLE_TYPE_ID = col_double()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   DAYS_TO_COLLECTION = col_double(),
#>   SAMPLE_INITIAL_WEIGHT = col_double(),
#>   OCT_EMBEDDED = col_logical(),
#>   SAMPLE_TYPE_ID = col_double()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   index = col_double(),
#>   chromosome = col_character(),
#>   region_start = col_double(),
#>   region_end = col_double(),
#>   peak_start = col_double(),
#>   peak_end = col_double(),
#>   enlarged_peak_start = col_double(),
#>   enlarged_peak_end = col_double(),
#>   n_genes_in_region = col_double(),
#>   genes_in_region = col_character(),
#>   n_genes_in_peak = col_double(),
#>   genes_in_peak = col_character(),
#>   n_genes_on_chip = col_character(),
#>   genes_on_chip = col_character(),
#>   `top 3` = col_character(),
#>   amp = col_double(),
#>   cytoband = col_character(),
#>   q_value = col_double()
#> )

#> Warning:  Multiple prefixes found, using keyword 'region' or taking first one

#> Parsed with column specification:
#> cols(
#>   index = col_double(),
#>   chromosome = col_double(),
#>   region_start = col_double(),
#>   region_end = col_double(),
#>   peak_start = col_double(),
#>   peak_end = col_double(),
#>   enlarged_peak_start = col_double(),
#>   enlarged_peak_end = col_double(),
#>   n_genes_in_region = col_double(),
#>   genes_in_region = col_character(),
#>   n_genes_in_peak = col_double(),
#>   genes_in_peak = col_character(),
#>   n_genes_on_chip = col_character(),
#>   genes_on_chip = col_character(),
#>   `top 3` = col_character(),
#>   amp = col_double(),
#>   cytoband = col_character(),
#>   q_value = col_double()
#> )

#> Warning:  Multiple prefixes found, using keyword 'region' or taking first one

#> A MultiAssayExperiment object of 11 listed
#>  experiments with user-defined names and respective classes.
#>  Containing an ExperimentList class object of length 11:
#>  [1] cna_hg19.seg: RaggedExperiment with 16080 rows and 90 columns
#>  [2] CNA: SummarizedExperiment with 24776 rows and 90 columns
#>  [3] linear_CNA: SummarizedExperiment with 24776 rows and 90 columns
#>  [4] methylation_hm450: SummarizedExperiment with 15473 rows and 80 columns
#>  [5] mutations_extended: RaggedExperiment with 20166 rows and 90 columns
#>  [6] mutations_mskcc: RaggedExperiment with 20166 rows and 90 columns
#>  [7] RNA_Seq_v2_expression_median: SummarizedExperiment with 20531 rows and 79 columns
#>  [8] RNA_Seq_v2_mRNA_median_all_sample_Zscores: SummarizedExperiment with 20531 rows and 79 columns
#>  [9] RNA_Seq_v2_mRNA_median_Zscores: SummarizedExperiment with 20440 rows and 79 columns
#>  [10] rppa_Zscores: SummarizedExperiment with 191 rows and 46 columns
#>  [11] rppa: SummarizedExperiment with 192 rows and 46 columns
#> Features:
#>  experiments() - obtain the ExperimentList instance
#>  colData() - the primary/phenotype DFrame
#>  sampleMap() - the sample availability DFrame
#>  `$`, `[`, `[[` - extract colData columns, subset, or experiment
#>  *Format() - convert into a long or wide DFrame
#>  assays() - convert ExperimentList to a SimpleList of matrices