A class for representing the cBioPortal API protocol

The `cBioPortal` class is a representation of the cBioPortal API protocol that directly inherits from the `Service` class in the `AnVIL` package. For more information, see the 'AnVIL' package.

Details

This class takes the static API as provided at https://www.cbioportal.org/api/api-docs and creates an R object with the help from underlying infrastructure (i.e., 'rapiclient' and 'AnVIL') to give the user a unified representation of the API specification provided by the cBioPortal group. Users are not expected to interact with this class other than to use it as input to the functionality provided by the rest of the package.

See also

Examples

cBioPortal()

#> service: cBioPortal
#> tags(); use cbioportal$<tab completion>:
#> # A tibble: 70 x 3
#>    tag             operation                          summary                   
#>    <chr>           <chr>                              <chr>                     
#>  1 A. Cancer Types getAllCancerTypesUsingGET          Get all cancer types      
#>  2 A. Cancer Types getCancerTypeUsingGET              Get a cancer type         
#>  3 B. Studies      fetchStudiesUsingPOST              Fetch studies by IDs      
#>  4 B. Studies      getAllStudiesUsingGET              Get all studies           
#>  5 B. Studies      getStudyUsingGET                   Get a study               
#>  6 B. Studies      getTagsForMultipleStudiesUsingPOST Get the study tags by IDs 
#>  7 B. Studies      getTagsUsingGET                    Get the tags of a study   
#>  8 C. Patients     fetchPatientsUsingPOST             fetchPatients             
#>  9 C. Patients     getAllPatientsInStudyUsingGET      Get all patients in a stu…
#> 10 C. Patients     getAllPatientsUsingGET             Get all patients          
#> # … with 60 more rows
#> tag values:
#>   A. Cancer Types, B. Studies, C. Patients, D. Samples, E. Sample
#>   Lists, F. Clinical Attributes, G. Clinical Data, H. Clinical Events,
#>   I. Molecular Data, J. Molecular Profiles, K. Mutations, L. Discrete
#>   Copy Number Alterations, M. Copy Number Segments, N. Genes, O. Gene
#>   Panels, P. Generic Assay, Q. Structural Variants, R. Reference Genome
#>   Genes, S. Resource Definitions, T. Resource Data
#> schemas():
#>   AlleleSpecificCopyNumber, CancerStudy, CancerStudyTags,
#>   ClinicalAttribute, ClinicalAttributeCount
#>   # ... with 51 more elements